Updating xcvrd test to work with media_settings_parser and moved the function check_port_in_range() from media_settings_parser back to xcvrd as optics_si_parser also uses it
